requirements.txt
astroid==2.4.1
autopep8==1.5.2
certifi==2020.4.5.1
chardet==3.0.4
checker==1.0
click==7.1.2
dbcm==1.0
flake8==3.8.2
Flask==1.1.2
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.770
mypy-extensions==0.4.3
mysql-connector-python==8.0.20
protobuf==3.12.1
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.5.2
requests==2.23.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
url-utils==1.0
urllib3==1.25.9
vsearch==1.0
Werkzeug==1.0.1
wrapt==1.12.1
